Refactor create, delete, update to move all queries to functions.php

+++ mysql/functions.php
@@ -1,6 +1,29 @@
 <?php include "db.php"; ?>
 <?php
 
+
+function createRows(){
+  if(isset($_POST['submit'])){
+    global $connection;
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+
+    $query = "INSERT INTO users(username, password) ";
+    $query .= "VALUES ('$username', '$password')";
+
+    $result = mysqli_query($connection, $query);
+
+    if(!$result) {
+      die('Query failed' . mysqli_error());
+    } else {
+      echo "Record created";
+    }
+
+  }
+}
+
+
+
 function showAllData() {
   global $connection;
   $query = "SELECT * FROM users";
@@ -19,33 +42,41 @@ function showAllData() {
 }
 
 function updateTable(){
-  global $connection;
-  $username = $_POST['username'];
-  $password = $_POST['password'];
-  $id = $_POST['id'];
+  if(isset($_POST['submit'])){
+    global $connection;
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+    $id = $_POST['id'];
 
-  $query = "UPDATE users SET ";
-  $query .= "username  = '$username', ";
-  $query .= "password = '$password' ";
-  $query .= "WHERE id = $id";
+    $query = "UPDATE users SET ";
+    $query .= "username  = '$username', ";
+    $query .= "password = '$password' ";
+    $query .= "WHERE id = $id";
 
-  $result = mysqli_query($connection, $query);
+    $result = mysqli_query($connection, $query);
 
-  if(!$result) {
-    die("QUERY FAILED " . mysqli_error($connection));
+    if(!$result) {
+      die("QUERY FAILED " . mysqli_error($connection));
+    } else {
+      echo "Record Updated";
+    }
   }
 }
 
 function deleteRows(){
-  global $connection;
-  $id = $_POST['id'];
+  if(isset($_POST['submit'])){
+    global $connection;
+    $id = $_POST['id'];
 
-  $query = "DELETE FROM users ";
-  $query .= "WHERE id = $id";
+    $query = "DELETE FROM users ";
+    $query .= "WHERE id = $id";
 
-  $result = mysqli_query($connection, $query);
+    $result = mysqli_query($connection, $query);
 
-  if(!$result) {
-    die("QUERY FAILED " . mysqli_error($connection));
+    if(!$result) {
+      die("QUERY FAILED " . mysqli_error($connection));
+    } else {
+      echo "Record deleted";
+    }
   }
 }
